import importStatus code to separate file
