+++ src/dashboard/dataImport/teamData.ts
@@ -1,5 +1,5 @@
 import { TournamentData } from 'schemas';
-import { setImportStatus } from '../globalScripts';
+import { setImportStatus } from '../importStatus';
 import { ImportStatus } from 'types/importStatus';
 import { sendLocalFile } from './postData';
