+++ src/dashboard/globalScripts.ts
@@ -1,7 +1,5 @@
 // For selectors
 
-import { ImportStatus } from 'types/importStatus';
-
 export function clearSelectors(className: string): void {
     const selectors = document.getElementsByClassName(className);
     for (let i = 0; i < selectors.length; i++) {
@@ -74,35 +72,3 @@ export function fillList(selectElem: HTMLSelectElement, data: Array<string>): vo
         selectElem.add(option);
     }
 }
-
-export function setImportStatus(status: ImportStatus, elem: HTMLElement): void {
-    let backgroundColor;
-    let textColor;
-    let text;
-    switch (status) {
-        case ImportStatus.Success:
-            backgroundColor = 'var(--green)';
-            textColor = 'white';
-            text = 'SUCCESS';
-            break;
-        case ImportStatus.Loading:
-            backgroundColor = 'var(--yellow)';
-            textColor = 'black';
-            text = 'LOADING';
-            break;
-        case ImportStatus.Failure:
-            backgroundColor = 'var(--red)';
-            textColor = 'white';
-            text = 'FAIL';
-            break;
-        default:
-            backgroundColor = '#181e29';
-            textColor = 'white';
-            text = 'UNKNOWN';
-    }
-
-    elem.style.backgroundColor = backgroundColor;
-    elem.style.color = textColor;
-    elem.innerText = text;
-}
-
